fix(poststat): fix argument -1 not accepted since 35a41fe21666708185078dc8cf2a89a1fc81ed29
This issue is caused by the comparison
between signed (`argc`; type: int)
and unsigned (`COUNTOF(myfile)`; type: size_t) integers.

This makes `argc < COUNTOF(myfile) - 1` always evaluates to false
for negative `argc`.

This fixes the issue by casting the unsigned integer (with type size_t)
into signed type `long`.

+++ util/poststat.c
@@ -465,7 +465,7 @@ main(
             break;
 
         default:
-            if (!(argc >= -1 && argc < COUNTOF(myfile) - 1))
+            if (!(argc >= -1 && argc < (long)COUNTOF(myfile) - 1))
             {
                 fprintf(stderr, "Usage: %s [action]\n", argv[0]);
                 fprintf(stderr, "actions:\n");
